+++ .changeset/hungry-peas-protect.md
@@ -1,5 +1,5 @@
 ---
-"@saleor/apps-shared": patch
+"@saleor/sentry-utils": minor
 ---
 
-Export release-utils module that helps with setting up release version for Sentry.
+Create sentry-utils module that will contain shared Sentry configuration. For now it exports release version that is used by Sentry during release.
