+++ apps/smtp/package.json
@@ -36,7 +36,7 @@
     "@saleor/apps-ui": "workspace:*",
     "@saleor/macaw-ui": "1.0.0-pre.7",
     "@saleor/react-hook-form-macaw": "workspace:*",
-    "@saleor/release-utils": "workspace:*",
+    "@saleor/sentry-utils": "workspace:*",
     "@sentry/nextjs": "8.0.0",
     "@tanstack/react-query": "4.29.19",
     "@trpc/client": "10.43.1",
