+++ packages/sentry-utils/.eslintrc
