+++ packages/sentry-utils/src/get-commit-hash.ts
