+++ packages/sentry-utils/src/get-release-tag.ts
